+++ modules/src/ics03_connection/msgs/conn_open_try.rs
@@ -166,6 +166,7 @@ impl TryFrom<RawMsgConnectionOpenTry> for MsgConnectionOpenTry {
                 msg.proof_init.into(),
                 client_proof,
                 Some(consensus_proof_obj),
+                None,
                 proof_height,
             )
             .map_err(|e| Kind::InvalidProof.context(e))?,