+++ modules/src/ics04_channel/msgs/acknowledgement.rs
@@ -68,6 +68,7 @@ impl TryFrom<RawMsgAcknowledgement> for MsgAcknowledgement {
             raw_msg.proof_acked.into(),
             None,
             None,
+            None,
             raw_msg
                 .proof_height
                 .ok_or(Kind::MissingHeight)?