+++ relayer/src/chain.rs
@@ -300,6 +300,7 @@ pub trait Chain: Sized {
                 CommitmentProofBytes::from(connection_proof),
                 client_proof,
                 consensus_proof,
+                None,
                 height.increment(),
             )
             .map_err(|_| Kind::MalformedProof)?,
@@ -328,8 +329,10 @@ pub trait Chain: Sized {
         let channel_proof =
             CommitmentProofBytes::from(self.proven_channel(port_id, channel_id, height)?.1);
 
-        Ok(Proofs::new(channel_proof, None, None, height.increment())
-            .map_err(|_| Kind::MalformedProof)?)
+        Ok(
+            Proofs::new(channel_proof, None, None, None, height.increment())
+                .map_err(|_| Kind::MalformedProof)?,
+        )
     }
 
     fn query_packet_commitments(
@@ -369,6 +372,8 @@ pub trait Chain: Sized {
         height: ICSHeight,
     ) -> Result<(Vec<u8>, Proofs), Error> {
         let data: Path;
+        let mut channel_proof = None;
+
         match packet_type {
             PacketMsgType::Recv => {
                 data = Path::Commitments {
@@ -391,6 +396,16 @@ pub trait Chain: Sized {
                     sequence,
                 }
             }
+            PacketMsgType::TimeoutOnClose => {
+                data = Path::Receipts {
+                    port_id: port_id.clone(),
+                    channel_id: channel_id.clone(),
+                    sequence,
+                };
+                channel_proof = Some(CommitmentProofBytes::from(
+                    self.proven_channel(port_id, channel_id, height)?.1,
+                ));
+            }
         }
 
         let res = self
@@ -403,6 +418,7 @@ pub trait Chain: Sized {
             })?),
             None,
             None,
+            channel_proof,
             height.increment(),
         )
         .map_err(|_| Kind::MalformedProof)?;