+++ relayer/src/chain/cosmos.rs
@@ -578,7 +578,7 @@ impl Chain for CosmosSDKChain {
         Ok((pc, height))
     }
 
-    /// Queries the packet commitment hashes associated with a channel.
+    /// Queries the unreceived packet sequences associated with a channel.
     fn query_unreceived_packets(
         &self,
         request: QueryUnreceivedPacketsRequest,
@@ -636,7 +636,7 @@ impl Chain for CosmosSDKChain {
         Ok((pc, height))
     }
 
-    /// Queries the packet commitment hashes associated with a channel.
+    /// Queries the unreceived acknowledgements sequences associated with a channel.
     fn query_unreceived_acknowledgements(
         &self,
         request: QueryUnreceivedAcksRequest,